Reference for the user-facing API

DataViewer.view — Function

view(data, name = "")

Open a viewer window for data, which may be a dictionary, HDF5 file, or possibly another key=>value based data type.

Optionally, name is a String indicating where data originates from (for example, a file name).

view(fname::String, ftype = "")

Open a viewer window for the data contained in file fname.

The extension of fname is normally used to determine how to read it. But the optional argument ftype may be provided to override this behavior. In this case, ftype should be a file extension associated to the file format (for example: ".jld2" for JLD2 files).

Because the file may need to be accessed during the whole browsing session, this function does not return until the window has been closed, at which point the data file itself is closed as well.

DataViewer.install — Function

install(; command::String = "dataviewer",
          destdir::String = joinpath(DEPOT_PATH[1], "bin"),
          force::Bool     = false,
          sysimage::Bool  = true)

Install a DataViewer launcher script to the destdir directory, with filename command. The default directory is a bin subfolder inside the first DEPOT_PATH entry, which usually translates to ~/.julia/bin. Add this directory to your PATH environment variable in order to more easily run the DataViewer launcher.

Set force to true to overwrite an existing destination file with the same path.

If sysimage is true (the default), a system image is compiled and the launcher script is set up to use it.